+++ jdbc-tracer-core/src/main/java/me/geso/jdbctracer/ResultSetListener.java
@@ -2,7 +2,6 @@ package me.geso.jdbctracer;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.List;
 
 public interface ResultSetListener {
     void trace(boolean first, ResultSet resultSet) throws SQLException;
