+++ jdbc-tracer-core/src/main/java/me/geso/jdbctracer/TracerConnection.java
