+++ jdbc-tracer-core/src/test/java/me/geso/jdbctracer/TracerConnectionTest.java
