+++ sample/build.gradle
@@ -1,10 +1,11 @@
+apply plugin:'java'
 apply plugin:'application'
 
 mainClassName = 'com.example.Main'
 applicationDefaultJvmArgs << '-Dorg.slf4j.simpleLogger.defaultLogLevel=debug'
 
 dependencies {
-    compile project(':jdbc-tracer')
+    compile project(':jdbc-tracer-driver')
     compile 'org.slf4j:slf4j-api:1.7.20'
     compile 'org.slf4j:slf4j-simple:1.7.20'
     compile 'org.projectlombok:lombok:1.16.8'
